优化协议编号可以是字符串类型

+++ noark-network/src/main/java/xyz/noark/network/codec/json/SimpleJsonCodec.java
@@ -46,7 +46,7 @@ public class SimpleJsonCodec extends AbstractPacketCodec {
 		ImmutableByteArray byteArray = new ImmutableByteArray(bytes.length + 4);
 		try (ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream(byteArray)) {
 			// 写入Opcode
-			byteArrayOutputStream.writeInt(networkProtocol.getOpcode());
+			byteArrayOutputStream.writeInt((Integer)networkProtocol.getOpcode());
 			// 写入协议内容
 			byteArrayOutputStream.writeBytes(bytes);
 		}
